# api-spec/api-spec.yaml
openapi: 3.0.0
info:
  version: "1.2.3"
  title: "Unicorn API"
  description: "API for managing unicorn."
servers:
  - url: https://api-stub-server-384446500375.asia-east1.run.app/
    description: "Backend server"
components:
...
  schemas:
    Account:
      type: object
      properties:
        uid:
          type: string
          description: "User ID"
          example: "1234567890"
        role:
          type: string
          description: "User Role"
          enum: ["user", "doctor"]
          example: "user"
        fcmTokenId:
          type: string
          description: "FCMトークンID"
          example: "fcm_token_id"
    AccountRequest:
      type: object
      properties:
        uid:
          type: string
          description: "User ID"
          example: "1234567890"
        role:
          type: string
          description: "User Role"
          enum: ["user", "doctor"]
          example: "user"
        fcmTokenId:
          type: string
          description: "FCMトークンID"
          example: "fcm_token_id"
    AccountResponse:
      type: object
      properties:
        uid:
          type: string
          description: "User ID"
          example: "1234567890"
        role:
          type: string
          description: "User Role"
          enum: ["user", "doctor"]
          example: "user"
        fcmTokenId:
          type: string
          description: "FCMトークンID"
          example: "fcm_token_id"
...
    Doctor:
      type: object
      properties:
        doctorID:
          type: string
          description: "医師ID"
          example: "1234567890"
        hospital:
          type: object
          properties:
            hospitalID:
              type: string
              description: "病院UUID"
              example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
            hospitalName:
              type: string
              description: "病院名"
              example: "東京都立大学病院"
        firstName:
          type: string
          description: "医師の名"
          example: "太郎"
        lastName:
          type: string
          description: "医師の姓"
          example: "山田"
        doctorIconUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        departments:
          type: array
          items:
            type: object
            properties:
              departmentName:
                type: string
                description: "診療科名"
                example: "内科"
              departmentID:
                type: string
                description: "診療科UUID"
                example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email:
          type: string
          description: "医師メールアドレス"
          example: "sample@mail.com"
        phoneNumber:
          type: string
          description: "医師電話番号"
          example: "09012345678"
        chatSupportHours:
          type: string
          description: "チャットサポート時間"
          example: "9:00-17:00"
        callSupportHours:
          type: string
          description: "通話サポート時間"
          example: "9:00-17:00"
    DoctorRequest:
      type: object
      properties:
        hospitalID:
          type: string
          description: "病院UUID"
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        firstName:
          type: string
          description: "医師の名"
          example: "太郎"
        lastName:
          type: string
          description: "医師の姓"
          example: "山田"
        doctorIconUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        departments:
          type: array
          items:
            type: string
            description: "診療科UUID"
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email:
          type: string
          description: "医師メールアドレス"
          example: "sample@mail.com"
        phoneNumber:
          type: string
          description: "医師電話番号"
          example: "09012345678"
        chatSupportStartHour:
          type: string
          description: "チャットサポート開始時間"
          example: "9:00"
        chatSupportEndHour:
          type: string
          description: "チャットサポート終了時間"
          example: "17:00"
        callSupportStartHour:
          type: string
          description: "通話サポート開始時間"
          example: "9:00"
        callSupportEndHour:
          type: string
          description: "通話サポート終了時間"
          example: "17:00"
    DoctorResponse:
      type: object
      properties:
        hospitalID:
          type: string
          description: "病院UUID"
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        doctorName:
          type: string
          description: "医師名"
          example: "山田 太郎"
        doctorIconUrl:
          type: string
          description: "画像URL"
          example: https://placehold.jp/150x150.png
        departments:
          type: array
          items:
            type: string
            description: "診療科UUID"
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email:
          type: string
          description: "医師メールアドレス"
          example: "sample@mail.com"
        phoneNumber:
          type: string
          description: "医師電話番号"
          example: "09012345678"
        chatSupportStartHour:
          type: string
          description: "チャットサポート開始時間"
          example: "9:00"
        chatSupportEndHour:
          type: string
          description: "チャットサポート終了時間"
          example: "17:00"
        callSupportStartHour:
          type: string
          description: "通話サポート開始時間"
          example: "9:00"
        callSupportEndHour:
          type: string
          description: "通話サポート終了時間"
          example: "17:00"
...
paths:
...
  /accounts:
    get:
      tags:
        - account
      summary: "アカウント情報取得"
      description: "アカウント情報を取得します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      responses:
        "200":
          description: "アカウント情報取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: "指定されたアカウントが存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "UserNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    post:
      tags:
        - account
      summary: "アカウント登録"
      description: "アカウントを登録します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRequest"
      responses:
        "200":
          description: "アカウント登録成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: "不正なデータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidData"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    delete:
      tags:
        - account
      summary: "アカウント削除"
      description: "アカウントを削除します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      responses:
        "204":
          description: "アカウント削除成功"
        "404":
          description: "指定されたアカウントが存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "UserNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
...
  /hospitals/{hospitalID}/doctors:
    get:
      tags:
        - doctor
      summary: "病院ごとの医師一覧取得"
      description: "病院の医師一覧を取得します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: hospitalID
          required: true
          schema:
            type: string
            description: "病院UUID"
      responses:
        "200":
          description: "医師一覧取得成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Doctor"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
  /doctors:
    get:
      tags:
        - doctor
      summary: "医師検索"
      description: "医者名、診療科、病院名で医師を検索します。全ての条件に一致する医師を返します（AND検索）。"
      parameters:
        - in: query
          name: doctorName
          schema:
            type: string
          required: false
          description: "医者名の部分一致検索"
        - in: query
          name: departmentID
          schema:
            type: string
          required: false
          description: "診療科IDで検索"
        - in: query
          name: hospitalName
          schema:
            type: string
          required: false
          description: "病院名の部分一致検索"
      responses:
        "200":
          description: "検索結果のリスト"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Doctor"
        "400":
          description: "不正なクエリパラメータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidQuery"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    post:
      tags:
        - doctor
      summary: "医師登録"
      description: "医師を登録します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRequest"
      responses:
        "200":
          description: "医師登録成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorResponse"
        "400":
          description: "不正なデータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidData"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
  /doctors/{doctorID}:
    get:
      tags:
        - doctor
      summary: "医師情報取得"
      description: "医師情報を取得します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: doctorID
          required: true
          schema:
            type: string
            description: "医師ID"
      responses:
        "200":
          description: "医師情報取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: "指定された医師が存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "DoctorNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
    put:
      tags:
        - doctor
      summary: "医師情報更新"
      description: "医師情報を更新します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: doctorID
          required: true
          schema:
            type: string
            description: "医師ID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRequest"
      responses:
        "200":
          description: "医師情報更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorResponse"
        "400":
          description: "不正なデータが送信されました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "InvalidData"
        "404":
          description: "指定された医師が存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "DoctorNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"

    delete:
      tags:
        - doctor
      summary: "医師削除"
      description: "医師を削除します"
      parameters:
        - in: header
          name: X-UID
          required: true
          schema:
            type: string
            description: "UID"
        - in: path
          name: doctorID
          required: true
          schema:
            type: string
            description: "医師ID"
      responses:
        "204":
          description: "医師削除成功"
        "404":
          description: "指定された医師が存在しません"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "DoctorNotFound"
        "500":
          description: "サーバーエラーが発生しました"
          content:
            application/json:
              schema:
                type: object
                properties:
                  errorType:
                    type: string
                    description: "エラータイプ"
                    example: "ServerError"
...